not to handle array with long etc...

+++ src/main/kotlin/com/github/ked4ma/atcoder/abc107_na/D.kt
@@ -1,8 +1,8 @@
 package com.github.ked4ma.atcoder.abc107_na
 
-import com.github.ked4ma.atcoder.utils.array.deprecated.*
+import com.github.ked4ma.atcoder.utils.array.any.d1.*
 import com.github.ked4ma.atcoder.utils.debug.*
-import com.github.ked4ma.atcoder.utils.input.deprecated.*
+import com.github.ked4ma.atcoder.utils.input.default.*
 import com.github.ked4ma.atcoder.utils.models.tree.fenwick.*
 
 fun main() {
@@ -22,14 +22,14 @@ fun main() {
             s[it + 1] = s[it] + if (data[it] >= x) 1 else -1
         }
         _debug_println(s.joinToString(", "))
-        val bit = FenwickTree(2 * N.toLong() + 1)
+        val bit = FenwickTree(2 * N + 1)
         var sum = 0L
         repeat(N + 1) {
             // calc inversion number.
             // target: S_i <= S_j
             // +1 means "=" of above fomula
-            sum += bit.sum(0, s[it] + N.toLong() + 1)
-            bit.add(s[it] + N.toLong(), 1) // count appeared number (S)
+            sum += bit.sum(0, s[it] + N + 1)
+            bit.add(s[it] + N, 1) // count appeared number (S)
         }
         _debug_println("$sum $m $x ${n / 2}")
         if (sum >= n / 2) {
