+++ src/main/kotlin/com/github/ked4ma/atcoder/abc330/E.kt
@@ -1,6 +1,5 @@
 package com.github.ked4ma.atcoder.abc330
 
-import com.github.ked4ma.atcoder.utils.array.any.*
 import com.github.ked4ma.atcoder.utils.array.any.d1.*
 import com.github.ked4ma.atcoder.utils.debug.*
 import com.github.ked4ma.atcoder.utils.input.default.*
@@ -11,7 +10,7 @@ import com.github.ked4ma.atcoder.utils.repeat.*
 fun main() {
     val (N, Q) = nextLongList()
     val A = nextLongList().toTypedArray()
-    val bk = sizedArray(N + 1, 0L)
+    val bk = sizedArray(N.toInt() + 1, 0L)
     A.forEach { a ->
         if (a <= N) {
             bk[a.toInt()]++
@@ -24,15 +23,15 @@ fun main() {
     _debug_println(set)
     loop(Q) {
         val (i, x) = nextLongList().let { (i, x) -> i - 1 to x }
-        if (A[i] <= N) {
-            bk[A[i]]--
-            if (bk[A[i]] == 0L) {
-                set.add(A[i])
+        if (A[i.toInt()] <= N) {
+            bk[A[i.toInt()].toInt()]--
+            if (bk[A[i.toInt()].toInt()] == 0L) {
+                set.add(A[i.toInt()])
             }
         }
-        A[i] = x
+        A[i.toInt()] = x
         if (x <= N) {
-            bk[x]++
+            bk[x.toInt()]++
             set.remove(x)
         }
         _debug_println(set)
