+++ src/main/kotlin/com/github/ked4ma/atcoder/abc332/E_x.kt
@@ -1,7 +1,5 @@
 package com.github.ked4ma.atcoder.abc332
 
-import com.github.ked4ma.atcoder.utils.array.any.*
-import com.github.ked4ma.atcoder.utils.array.double.*
 import com.github.ked4ma.atcoder.utils.array.double.d2.*
 import com.github.ked4ma.atcoder.utils.debug.*
 import com.github.ked4ma.atcoder.utils.input.default.*
@@ -15,16 +13,16 @@ fun main() {
     val W = nextLongList()
     val avg = W.sum().toDouble() / D
     _debug_println(avg)
-    val dp = sized2DDoubleArray(1L shl N.toInt(), D + 1, 0.0)
-    for (i in 0 until (1L shl N.toInt())) {
+    val dp = sized2DDoubleArray(1 shl N.toInt(), D.toInt() + 1, 0.0)
+    for (i in 0 until (1 shl N.toInt())) {
         var y = 0L
-        for (j in 0 until N) {
-            if (i and (1L shl j.toInt()) > 0L) {
+        for (j in 0 until N.toInt()) {
+            if (i and (1 shl j) > 0L) {
                 y += W[j]
             }
         }
         dp[i][1] = (y - avg).pow(2)
-        for (k in 2..D) {
+        for (k in 2..D.toInt()) {
             dp[i][k] = dp[i][k - 1] + dp[0][1]
             var x = i
             while (x > 0) {
@@ -34,5 +32,5 @@ fun main() {
             }
         }
     }
-    println("%.10f".format(dp[(1L shl N.toInt()) - 1][D] / D))
+    println("%.10f".format(dp[(1 shl N.toInt()) - 1][D.toInt()] / D))
 }
